AAE-47794 Add RuntimeBundleShedLockConfiguration for ShedLock integration
